Added content, BasicLinux.html, BasicMaven.html
Added BasicLinux.html to house useful tips and tricks
Changed maven.html to BasicMaven.html to better reflect the level of the
information therein and to ensure conformity in file naming.
Added content to BasicMaven.html, including information and links
about testing, resources, and how to add a third-party local jar as a
dependency.